Reformatted main drawing and updates scene image
Reformatted the layout of the DrawSceneMap to better handle events and
cuts down on unnecessary work. It appears slightly longer but less is
done per call. It may be worth changing again so there's a switch inside
a loop instead of 3 loops depending on the event Type.

When a scene loads it will update the image for that scene.

It now better manages the textures and stops the leak when a scene
loads.