Disable before unloading, calculates closest point
Added a singleton for SceneManager for ManagedSceneObject to register
to.

For testing reasons I used a circle to check if the player was in range
of a scene to try to load it. I never actually got round to testing it
and it obviously didn't work (should have x1.4 at least), instead it now
finds the closest point to the scene rect from the tracked object
(camera) and measures that distance. Ideally this should be changed from
closest point of both Camera rect (clip planes) and scene rect.

There are now 3 states a scene can be in. Loaded and active, loaded and
inactive and not loaded. An extra state was added incase the player
backtracks so the whole thing won't have to be loaded again.