OOR-index bugfix, instance set, rect draw change
If the last object in the asset list was select then removed, it would
cause an out-of-range exception on redraw as the selected and hovered
indices weren't reset.

the window instance was only being set on creation, not OnEnable so it
wasn't being properly set again on reload.

Changed the dotted line in the scene view to be more easily visible on
backgrounds,